+++ src/main/main.go
@@ -1,9 +1,9 @@
 package main
 
 import (
-	"db"
 	"github.com/DeKugelschieber/go-session"
-	"log"
+	log "github.com/sirupsen/logrus"
+	"model"
 	"net/http"
 	"os"
 	"path/filepath"
@@ -47,7 +47,7 @@ func logToFile() *os.File {
 
 // Starts the RESTful server.
 func startServer() {
-	log.Print("Starting server on ", os.Getenv("ACWEB_HOST"))
+	log.Info("Starting server on ", os.Getenv("ACWEB_HOST"))
 
 	mux := http.NewServeMux()
 	mux.Handle("/robots.txt", http.HandlerFunc(returnRobotsTxt))
@@ -96,8 +96,8 @@ func main() {
 	}
 
 	// connect to db
-	db.Connect(os.Getenv("ACWEB_DB_USER"), os.Getenv("ACWEB_DB_PASSWORD"), os.Getenv("ACWEB_DB_HOST"), os.Getenv("ACWEB_DB"))
-	defer db.Disconnect()
+	model.Connect(os.Getenv("ACWEB_DB_USER"), os.Getenv("ACWEB_DB_PASSWORD"), os.Getenv("ACWEB_DB_HOST"), os.Getenv("ACWEB_DB"))
+	defer model.Disconnect()
 
 	// start session manager
 	sessionProvider := session.NewMemProvider()