+++ src/rest/api.go
@@ -5,8 +5,8 @@ import (
 	"encoding/json"
 	"github.com/DeKugelschieber/go-resp"
 	"github.com/DeKugelschieber/go-session"
+	log "github.com/sirupsen/logrus"
 	"instance"
-	"log"
 	"model"
 	"net/http"
 	"settings"
@@ -77,12 +77,12 @@ func CheckSession(w http.ResponseWriter, r *http.Request) {
 		var id int64
 
 		if err := s.Get("user_id", &id); err != nil {
-			log.Printf("Error reading user ID: %v", err)
+			log.WithFields(log.Fields{"err": err}).Error("Error reading user ID")
 			resp.Error(w, 1, "Error reading user ID", nil)
 			return
 		}
 
-		resp.Success(w, 0, "", loginRes{id})
+		resp.Success(w, 0, "", struct{Id int64 `json:"user_id"`}{id})
 	} else {
 		// don't log this
 		resp.Log = false
@@ -92,7 +92,10 @@ func CheckSession(w http.ResponseWriter, r *http.Request) {
 }
 
 func Login(w http.ResponseWriter, r *http.Request) {
-	var req loginReq
+	req := struct {
+		Login string `json:"login"`
+	Pwd   string `json:"pwd"`
+	}{}
 
 	if decode(w, r, &req) {
 		return
@@ -108,7 +111,7 @@ func Login(w http.ResponseWriter, r *http.Request) {
 	s, err := session.NewSession(w, r)
 
 	if err != nil {
-		log.Printf("Error starting session on login: %v", err)
+		log.WithFields(log.Fields{"err": err}).Error("Error starting session on login")
 		resp.Error(w, 3, err.Error(), nil)
 		return
 	}
@@ -118,7 +121,7 @@ func Login(w http.ResponseWriter, r *http.Request) {
 	s.Set("moderator", user.Moderator)
 
 	if err := s.Save(); err != nil {
-		log.Printf("Error saving session on login: %v", err)
+		log.WithFields(log.Fields{"err": err}).Error("Error saving session on login")
 		resp.Error(w, 4, err.Error(), nil)
 		return
 	}
@@ -130,13 +133,13 @@ func Logout(w http.ResponseWriter, r *http.Request) {
 	s, err := session.GetCurrentSession(r)
 
 	if !s.Active() {
-		log.Printf("Session not found on logout: %v", err)
+		log.WithFields(log.Fields{"err": err}).Error("Session not found on logout")
 		resp.Error(w, 1, "Session not found", err)
 		return
 	}
 
 	if err := s.Destroy(w, r); err != nil {
-		log.Printf("Error destroying user session on logout: %v", err)
+		log.WithFields(log.Fields{"err": err}).Error("Error destroying user session on logout")
 		resp.Error(w, 2, "Error destroying user session", nil)
 		return
 	}
@@ -150,7 +153,15 @@ func AddEditUser(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	var req addEditUserReq
+	req := struct {
+		Id        int64  `json:"id"`
+		Login     string `json:"login"`
+		Email     string `json:"email"`
+		Pwd1      string `json:"pwd1"`
+		Pwd2      string `json:"pwd2"`
+		Admin     bool   `json:"admin"`
+		Moderator bool   `json:"moderator"`
+	}{}
 
 	if decode(w, r, &req) {
 		return
@@ -238,7 +249,11 @@ func SaveSettings(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	var req saveSettingsReq
+	req := struct {
+		Folder     string `json:"folder"`
+		Executable string `json:"executable"`
+		Args       string `json:"args"`
+	}{}
 
 	if decode(w, r, &req) {
 		return
@@ -359,7 +374,10 @@ func StartInstance(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	var req startInstanceReq
+	req := struct {
+		Name          string `json:"name"`
+		Configuration int64  `json:"config"`
+	}{}
 
 	if decode(w, r, &req) {
 		return