+++ src/rest/settingsTypes.go
@@ -1,7 +1,0 @@
-package rest
-
-type saveSettingsReq struct {
-	Folder     string `json:"folder"`
-	Executable string `json:"executable"`
-	Args       string `json:"args"`
-}