+++ src/user/get.go
@@ -2,7 +2,7 @@ package user
 
 import (
 	"github.com/DeKugelschieber/go-util"
-	"log"
+	log "github.com/sirupsen/logrus"
 	"model"
 )
 
@@ -10,7 +10,7 @@ func GetAllUser() ([]model.User, error) {
 	user, err := model.GetAllUser()
 
 	if err != nil {
-		log.Printf("Error reading all user: %v", err)
+		log.WithFields(log.Fields{"err": err}).Error("Error reading all user")
 		return nil, util.OpError{1, "Error reading all user"}
 	}
 
@@ -21,7 +21,7 @@ func GetUser(id int64) (*model.User, error) {
 	user, err := model.GetUserById(id)
 
 	if err != nil {
-		log.Printf("Error reading user by ID: %v", err)
+		log.WithFields(log.Fields{"err": err}).Error("Error reading user by ID")
 		return nil, util.OpError{1, "Error reading user"}
 	}