+++ src/user/remove.go
@@ -2,7 +2,7 @@ package user
 
 import (
 	"github.com/DeKugelschieber/go-util"
-	"log"
+	log "github.com/sirupsen/logrus"
 	"model"
 )
 
@@ -10,12 +10,12 @@ func RemoveUser(id int64) error {
 	user, err := model.GetUserById(id)
 
 	if err != nil {
-		log.Printf("Error reading user by ID: %v", err)
+		log.WithFields(log.Fields{"err": err}).Error("Error reading user by ID")
 		return util.OpError{1, "Error reading user"}
 	}
 
 	if err := user.Remove(); err != nil {
-		log.Printf("Error removing user: %v", err)
+		log.WithFields(log.Fields{"err": err}).Error("Error removing user")
 		return util.OpError{2, "Error removing user"}
 	}