+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.0.4
+
+* fixed: some default config values
+* added changing dynamic track conditions
+
 ## 1.0.3
 
 Now providing a docker image! Configuration is now done via environment variables, see README.md for details.